Add macro flag to identify triggering entity
At present it is difficult to identify within a macro which entity triggered it. This makes reusing macros for recurring triggers difficult because they require either duplicate copies for each trigger or detection workarounds. This change simplifies the process by creating a time-limited flag on the `Macro` itself identifying the entity type and its id, which can be accessed from within the macro.